capstone-project/requirements.txt
agate==1.6.3
Babel==2.11.0
boto3==1.26.32
botocore==1.29.32
click==8.1.3
colorama==0.4.5
dbt-core==1.3.1
dbt-extractor==0.4.1
dbt-postgres==1.3.1
dbt-redshift==1.3.0
future==0.18.2
hologram==0.0.15
isodate==0.6.1
jmespath==1.0.1
jsonschema==3.2.0
leather==0.3.4
Logbook==1.5.3
mashumaro==3.0.4
minimal-snowplow-tracker==0.0.2
msgpack==1.0.4
networkx==2.8.8
parsedatetime==2.4
pathspec==0.9.0
psycopg2-binary==2.9.5
python-slugify==7.0.0
pytimeparse==1.1.8
pytz==2022.6
s3transfer==0.6.0
sqlparse==0.4.3
text-unidecode==1.3
Werkzeug==2.2.2